move legacy check to assets namespace
