+++ wp-content/plugins/core/src/Theme/Theme_Subscriber.php
@@ -4,6 +4,7 @@ declare( strict_types=1 );
 namespace Tribe\Project\Theme;
 
 use Tribe\Libs\Container\Abstract_Subscriber;
+use Tribe\Project\Assets\Theme\Legacy_Check;
 use Tribe\Project\Theme\Config\Image_Sizes;
 use Tribe\Project\Theme\Config\Supports;
 use Tribe\Project\Theme\Config\Web_Fonts;
@@ -13,7 +14,6 @@ use Tribe\Project\Theme\Media\Full_Size_Gif;
 use Tribe\Project\Theme\Media\Image_Wrap;
 use Tribe\Project\Theme\Media\Oembed_Filter;
 use Tribe\Project\Theme\Media\WP_Responsive_Image_Disabler;
-use Tribe\Project\Theme\Resources\Legacy_Check;
 use Tribe\Project\Theme\Resources\Login_Resources;
 
 class Theme_Subscriber extends Abstract_Subscriber {
@@ -23,7 +23,6 @@ class Theme_Subscriber extends Abstract_Subscriber {
 		$this->config();
 
 		$this->login_resources();
-		$this->legacy_resources();
 
 		$this->editor();
 	}
@@ -108,20 +107,6 @@ class Theme_Subscriber extends Abstract_Subscriber {
 		add_action( 'login_enqueue_scripts', function () {
 			$this->container->get( Login_Resources::class )->login_styles();
 		}, 10, 0 );
-	}
-
-	private function legacy_resources() {
-		add_action( 'wp_head', function () {
-			$this->container->get( Legacy_Check::class )->old_browsers();
-		}, 0, 0 );
-
-		add_action( 'init', function () {
-			$this->container->get( Legacy_Check::class )->add_unsupported_rewrite();
-		} );
-
-		add_filter( 'template_include', function ( $template ) {
-			return $this->container->get( Legacy_Check::class )->load_unsupported_template( $template );
-		} );
 	}
 
 	private function web_fonts() {
